feat: extracting hooks logic
